Break out encode and decode into their own files
