+++ mod.ts
@@ -1,4 +1,6 @@
 export * from "./src/CBOR.ts";
+export * from "./src/decode.ts";
+export * from "./src/encode.ts";
 export * from "./src/SimpleValue.ts";
 export * from "./src/TaggedValue.ts";
 export * from "./src/types.ts";
